refactor: 각 기능별로 할 수 있게 파일 구조 및 폴더 구조 변경
각 sub, sidebar, nav, 공통 기능(all)로 나눠서 폴더 구조를 변경하였으며, 파일도 좀 더 세분화하게 변경하였습니다.
