--- css/font.css
@@ -1,4 +1,0 @@
-@font-face {
-    font-family: ‘Pretendard’;
-    src: url(/fonts/Pretendard-Black.ttf);
-}
