--- js/all_js.js
@@ -1,6 +1,0 @@
-import "/js/dimmed/navDimmed.js";
-import "/js/dimmed/sidebarDimmed.js";
-import "/js/sidebar/showSidebar.js";
-import "/js/sidebar/sidebarInputData.js";
-import "/js/allQuery.js";
-import "/model/data.js";
