+++ main.html
@@ -1,9 +1,9 @@
 <!DOCTYPE html>
-<html lang="en">
+<html lang="ko">
 <head>
     <meta charset="UTF-8">
-    <title>Title</title>
-    <link rel="stylesheet" href="css/all-css.css">
+    <title>mainpage</title>
+    <link rel="stylesheet" href="styleSheet/all/allCss.css">
 </head>
 <body>
 <header>
@@ -73,30 +73,43 @@
                             <a href="#" class="login_popup_text_click">
                                 <div>여기에서 시작합니다.</div></a>
                         </div>
-                        <div class="nav__login_popup_extension">
+                    </div>
+                </div>
+                //
+                <div class="nav__login_popup_extension">
+                    <div class="nav__login_popup-shape">
+                    </div>
+                    <div class="nav__login_popup-box_extension">
+                        <button class="nav__login_popup_button">로그인</button>
+                        <div class="login_popup_text">
+                            <div>기존 사용자가 아니십니까?</div>
+                            <a href="#" class="login_popup_text_click">
+                                <div>여기에서 시작합니다.</div></a>
+                        </div>
+                        <div class="nav__login_popup_extension_list">
                             <hr>
                             <div class="login_popup_extension-left-box">
                                 <ul>
                                     <li class="login_popup_extension-title">귀하의 목록</li>
-                                    <li class="login_popup_extension-list">목록 생성</li>
-                                    <li class="login_popup_extension-list">목록 또는 레지스트리 찾기</li>
-                                    <li class="login_popup_extension-list">AmazonSmile 자선 품목 목록</li>
+                                    <li>목록 생성</li>
+                                    <li>목록 또는 레지스트리 찾기</li>
+                                    <li>AmazonSmile 자선 품목 목록</li>
                                 </ul>
                             </div>
                             <div class="login_popup_extension-right-box">
                                 <ul>
                                     <li class="login_popup_extension-title">계정</li>
-                                    <li class="login_popup_extension-list">계정</li>
-                                    <li class="login_popup_extension-list">주문</li>
-                                    <li class="login_popup_extension-list">권장 사항</li>
-                                    <li class="login_popup_extension-list">검색 기록</li>
-                                    <li class="login_popup_extension-list">위치리스트</li>
-                                    <li class="login_popup_extension-list">비디오 구매 및 대여</li>
-                                    <li class="login_popup_extension-list">Kindle 언리미티드</li>
-                                    <li class="login_popup_extension-list">콘테츠 및 기기</li>
-                                    <li class="login_popup_extension-list">항목 구독 및 저장</li>
-                                    <li class="login_popup_extension-list">멤버십 및 구독</li>
-                                    <li class="login_popup_extension-list">음악 라이브러리</li>
+                                    <li>계정</li>
+                                    <li>주문</li>
+                                    <li>권장 사항</li>
+                                    <li>검색 기록</li>
+                                    <li>위치리스트</li>
+                                    <li>비디오 구매 및 대여</li>
+                                    <li>Kindle 언리미티드</li>
+                                    <li>콘테츠 및 기기</li>
+                                    <li>항목 구독 및 저장</li>
+                                    <li>멤버십 및 구독</li>
+                                    <li>음악 라이브러리</li>
                                 </ul>
                             </div>
                         </div>
@@ -117,15 +130,15 @@
             </div>
         </div>
         <ul class="sub flex__row">
-            <li ><a href="#" class="sub__all-menu align__center sub__menu">
+            <li ><div class="sub__all-menu align__center">
                 <img src="assets/menu.svg">
                 <div >모두</div>
-            </a></li>
-            <li class="sub__menu">오늘의 딜</li>
-            <li class="sub__menu">고객 서비스</li>
-            <li class="sub__menu">레지스트리</li>
-            <li class="sub__menu">기프트 카드</li>
-            <li class="sub__menu">판매</li>
+            </div></li>
+            <li>오늘의 딜</li>
+            <li>고객 서비스</li>
+            <li>레지스트리</li>
+            <li>기프트 카드</li>
+            <li>판매</li>
             <li class="sub__shopping">지금 특가 상품 쇼핑하기</li>
         </ul>
     </div>
@@ -189,5 +202,5 @@
 
 
 </body>
-<script defer type="module" src="js/all_js.js" ></script>
+<script defer type="module" src="scripts/all/all_js.js" ></script>
 </html>
