+++ scripts/all/allJs.js
@@ -1,0 +1,6 @@
+import "/Users/an-yelim/codesquad/fe--amazon/scripts/nav/navDimmed.js";
+import "/Users/an-yelim/codesquad/fe--amazon/scripts/sidebar/sidebarDimmed.js";
+import "/Users/an-yelim/codesquad/fe--amazon/scripts/sidebar/showSidebar.js";
+import "/Users/an-yelim/codesquad/fe--amazon/scripts/sidebar/sidebarInputData.js";
+import "/Users/an-yelim/codesquad/fe--amazon/scripts/all/allQuery.js";
+import "/model/data.js";
