+++ scripts/all/allQuery.js
