+++ styleSheet/nav/navAnimation.css
@@ -1,0 +1,21 @@
+@keyframes slowshow {
+    from{
+        opacity: 0;
+        visibility: hidden;
+    }
+    to{
+        opacity: 1;
+        visibility: visible;
+         }
+}
+
+.nav__login_popup{
+    visibility: hidden;
+    position: absolute;
+    width: 235px;
+    height: 100px;
+    left: -170px;
+    top: 44px;
+    animation: slowshow 1s forwards;
+    animation-delay: 1s;
+}
